Aula11-03

+++ src/main/java/br.com.fiap.api_rest/controller/ClienteController.java
@@ -5,6 +5,12 @@ import br.com.fiap.api_rest.dto.ClientResponse;
 import br.com.fiap.api_rest.model.Cliente;
 import br.com.fiap.api_rest.repository.ClienteRepository;
 import br.com.fiap.api_rest.service.ClientService;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.media.Content;
+import io.swagger.v3.oas.annotations.media.Schema;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.responses.ApiResponses;
+import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
@@ -19,34 +25,65 @@ import java.util.List;
 import java.util.Optional;
 
 @RestController
-@RequestMapping("/clientes")
+@RequestMapping(value="/clientes",produces = {"application/json"})
+@Tag(name = "api-clientes")
 public class ClienteController {
     @Autowired
     ClienteRepository clienteRepository;
     @Autowired
-    ClientService clientService;
-    // Create, Read, Update, Delete
-    // Post, Get, Put, Delete
+    ClientService clienteService;
 
+    // Create, Read, Update, Delete - CRUD
+    // Post, Get, Put, Delete - Verbos HTTP correspondentes
+    @Operation(summary = "Cria um novo cliente")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "201",
+                    description = "Cliente cadastrado com sucesso",
+                    content = {@Content(mediaType = "application/json",schema = @Schema(implementation = Cliente.class))}),
+            @ApiResponse(responseCode = "400",
+                    description = "Atributos inválidos",
+                    content = {@Content(schema = @Schema())})
+
+    })
     @PostMapping
     public ResponseEntity<Cliente> createCliente(@Valid @RequestBody ClientRequest cliente) {
-        Cliente clienteSalvo = clienteRepository.save(clientService.requestToCliente(cliente));
+        Cliente clienteSalvo = clienteRepository.save(clienteService.requestToCliente(cliente));
         return new ResponseEntity<>(clienteSalvo, HttpStatus.CREATED);
     }
-
+    @Operation(summary = "Retorna uma lista de clientes")
     @GetMapping
-    public ResponseEntity<Page<ClientResponse>> readClientes(@RequestParam(required = true) int page) {
-        Pageable  pageable = PageRequest.of( page,2, Sort.by("nome").ascending().and(Sort.by("categoria")));
-        return new ResponseEntity<>(clientService.findAll(pageable), HttpStatus.OK);
+    public ResponseEntity<Page<ClientResponse>> readClientes(@RequestParam(defaultValue = "0") Integer page) {
+        Pageable pageable = PageRequest.of(page, 2, Sort.by("categoria").ascending().and(Sort.by("nome").ascending()));
+        return new ResponseEntity<>(clienteService.findAll(pageable), HttpStatus.OK);
     }
-    //PathVariable parametro na url ex: /clientes/1
-    //Requestparam parametro como query ex: /clientes/?id=1
+    // PathVariable = parâmetro diretamente na URL, ex: /clientes/1
+    // RequestParam = parâmetro como query, ex: /clientes/?id=1
+    @Operation(summary = "Retorna um cliente por id")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200",
+                    description = "Cliente encontrado com sucesso",
+                    content = {@Content(mediaType = "application/json",schema = @Schema(implementation = ClientResponse.class))}),
+            @ApiResponse(responseCode = "404",
+                    description = "Nenhum cliente encontrado",
+                    content = {@Content(schema = @Schema())})
+
+    })
     @GetMapping("/{id}")
     public ResponseEntity<ClientResponse> readCliente(@PathVariable Long id) {
         Optional<Cliente> cliente = clienteRepository.findById(id);
-        return cliente.map(value -> new ResponseEntity<>(clientService.clientToResponse(value), HttpStatus.OK)).orElseGet(() -> new ResponseEntity<>(HttpStatus.NOT_FOUND));
+        return cliente.map(value -> new ResponseEntity<>(clienteService.clientToResponse(value, false), HttpStatus.OK))
+                .orElseGet(() -> new ResponseEntity<>(HttpStatus.NOT_FOUND));
     }
+    @Operation(summary = "Atualiza um cliente existente")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "201",
+                    description = "Cliente encontrado e atualizado com sucesso",
+                    content = {@Content(mediaType = "application/json",schema = @Schema(implementation = Cliente.class))}),
+            @ApiResponse(responseCode = "400",
+                    description = "Nenhum cliente encontrado para atualizar",
+                    content = {@Content(schema = @Schema())})
 
+    })
     @PutMapping("/{id}")
     public ResponseEntity<Cliente> updateCliente(@PathVariable Long id,
                                                  @RequestBody Cliente cliente) {
@@ -58,9 +95,22 @@ public class ClienteController {
         Cliente clienteAtualizado = clienteRepository.save(cliente);
         return new ResponseEntity<>(clienteAtualizado, HttpStatus.CREATED);
     }
+    @Operation(summary = "Exclui um cliente por id")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "204",
+                    description = "Cliente encontrado e excluido com sucesso",
+                    content = {@Content(schema = @Schema())}),
+            @ApiResponse(responseCode = "400",
+                    description = "Nenhum cliente encontrado para excluir",
+                    content = {@Content(schema = @Schema())})
 
+    })
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> deleteCliente(@PathVariable Long id) {
+        Optional<Cliente> clienteExistente = clienteRepository.findById(id);
+        if (clienteExistente.isEmpty()) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
         clienteRepository.deleteById(id);
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
